\name{CNVtest-package}
\alias{CNVtest-package}
\docType{package}
\title{
Identify both Sample-Specific CNVs and Trait-Associated CNVs 
}
\description{
This package is developed to directly identify the copy number variants (CNVs) that are associated with outcome of interest, without the need of identifying sample-specific CNVs (e.g. GWAS). 

In addition, this package includes optimal likelihood ratio selection (LRS) procedure, which detect the CNV regions for one sample with optimal rate. 

}
\details{
\tabular{ll}{
Package: \tab CNVtest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-07\cr
License: \tab GPL (>= 2  \cr
LazyLoad: \tab yes \cr
}
}
\author{
Author: Vicky Qian Wu ; X. Jessie Jeng ; Hongzhe Li <hongzhe@upenn.edu>


Maintainer: Vicky Qian Wu <wuqian7@gmail.com>
}
\references{
X. Jessie Jeng, Qian Wu and Hongzhe Li (2015) A Statistical Method for Identifying Trait-Associated Copy Number Variants. \eqn{Human heredity};79:147-156

X. Jessie Jeng, T. Tony Cai, and Hongzhe Li (2010). Optimal Sparse Segment Identification with Application in Copy Number Variation Analysis. \eqn{ Journal of American Statistical Association }, T&M. 105 (491), 1156-1166.
}

\keyword{ package }

