\name{Example_pick}
\alias{Example_pick}
\docType{data}
\title{
Data of One Sample CNV region
}
\description{
Example of pick the CNV region for one sample 
}
\usage{Example_pick}

\references{
X. Jessie Jeng, T. Tony Cai, and Hongzhe Li (2010). Optimal Sparse Segment Identification with Application in Copy Number Variation Analysis. \eqn{Journal of American Statistical Association}, T&M. 105 (491), 1156-1166. 
}

\examples{
data(Example_pick)
}
\keyword{datasets}
